+++ src/api.ts
@@ -2,11 +2,9 @@
 import { camelizeKeys } from 'humps';
 import { NextApiRequest } from 'next';
 
-import { logError } from '@/lib/newrelic';
 import { MushafLines, QuranFont } from '@/types/QuranReader';
 import { SearchRequestParams, SearchMode } from '@/types/Search/SearchRequestParams';
 import NewSearchResponse from '@/types/Search/SearchResponse';
-import { getDefaultWordFields, getMushafId } from '@/utils/api';
 import {
   makeAdvancedCopyUrl,
   makeTafsirsUrl,
@@ -33,11 +31,7 @@ import {
   makeCountryLanguagePreferenceUrl,
 } from '@/utils/apiPaths';
 import { getAdditionalHeaders } from '@/utils/headers';
-import {
-  AdvancedCopyRequest,
-  PagesLookUpRequest,
-  QuranInYearVerseRequest,
-} from 'types/ApiRequests';
+import { AdvancedCopyRequest, PagesLookUpRequest } from 'types/ApiRequests';
 import {
   TranslationsResponse,
   AdvancedCopyRawResultResponse,
@@ -94,46 +88,6 @@ export const fetcher = async function fetcher<T>(
   }
   const json = await res.json();
   return camelizeKeys(json);
-};
-
-/**
- * Fetch the verse for the current "Quran in a Year" day with proper word fields and mushaf config.
- *
- * @param {QuranInYearVerseRequest} params - Request parameters
- * @param {string} params.locale - Target locale for the request
- * @param {number} params.chapter - Chapter number of the target ayah
- * @param {number} params.verse - Verse number of the target ayah
- * @param {number[]} params.translationIds - Preferred translation IDs, first one is used; empty means no translations
- * @param {MushafLines} params.mushafLines - Mushaf line configuration to build the mushaf id
- * @returns {Promise<VersesResponse | undefined>} Verse payload or undefined on failure
- */
-export const getQuranInYearVerse = async ({
-  locale,
-  chapter,
-  verse,
-  translationIds,
-  mushafLines,
-}: QuranInYearVerseRequest): Promise<VersesResponse | undefined> => {
-  try {
-    const translationsParam = translationIds.slice(0, 1).join(',');
-    const quranInYearParams = {
-      ...getDefaultWordFields(QuranFont.QPCHafs),
-      translationFields: 'resource_name,language_id',
-      ...(translationsParam ? { translations: translationsParam } : {}),
-      mushaf: getMushafId(QuranFont.QPCHafs, mushafLines).mushaf,
-      from: `${chapter}:${verse}`,
-      to: `${chapter}:${verse}`,
-    };
-
-    return fetcher<VersesResponse>(makeVersesUrl(chapter, locale, quranInYearParams));
-  } catch (error) {
-    const err = error instanceof Error ? error : new Error(String(error));
-    logError('Failed to fetch Quran in a Year verse', err, {
-      locale,
-      from: `${chapter}:${verse}`,
-    });
-    return undefined;
-  }
 };
 
 /**
